refactor: split interfaces into separate files
