+++ src/app/models/attributeTypeDictionary.ts
@@ -1,0 +1,5 @@
+import { AttributeType } from './attributeType';
+
+export interface AttributeTypeDictionary {
+  [key: string]: AttributeType;
+}
