+++ src/app/models/candidateAttribute.ts
@@ -1,0 +1,9 @@
+import { CandidateAttributeType } from './candidateAttributeType';
+
+export interface CandidateAttribute {
+  id: number;
+  value: string;
+  valueSource: number;
+  isArchived: boolean;
+  attributeTypes: CandidateAttributeType;
+}
