+++ src/app/models/candidateAttributeType.ts
@@ -1,0 +1,7 @@
+export interface CandidateAttributeType {
+  id: number;
+  name: string;
+  basicType: string;
+  validation: string;
+  identifier: boolean;
+}
