+++ src/app/models/candidateAttributesValues.ts
@@ -1,0 +1,4 @@
+export interface CandidateAttributesValues {
+  name: string;
+  value: string;
+}
