+++ src/app/models/candidatesFilterData.ts
@@ -1,0 +1,5 @@
+export interface CandidatesFilterData {
+  pageNumber?: number;
+  pageSize?: number;
+  keyword?: string[];
+}
