+++ src/app/models/commentData.ts
@@ -1,0 +1,5 @@
+export interface CommentData {
+  comment: string;
+  id?: number;
+  archived?: boolean;
+}
