+++ src/app/models/communicationHistory.ts
@@ -1,0 +1,7 @@
+export interface CommunicationHistory {
+  id: number;
+  createDate: string;
+  updateDate: string;
+  comment: string;
+  archived: boolean;
+}
