+++ src/app/models/keywords.ts
@@ -1,0 +1,8 @@
+/**
+ * @typedef Keywords
+ * @prop {string} id Id
+ */
+
+export interface Keywords {
+  id: string;
+}
