+++ src/app/utils/functions.ts
@@ -3,7 +3,7 @@ import { NotificationService } from '@services/notification.service';
 import { ERROR_STATUS_CODES } from '@constants/errorStatusCode';
 import { NotificationMode } from '@constants/notification';
 import { ERROR_MESSAGE } from '@constants/strings';
-import { CandidateAttribute } from '@src/app/models/candidates';
+import { CandidateAttribute } from '../models/candidateAttribute';
 
 export const defaultErrorhandler = (
   notification: NotificationService,
